reorg shim coverage tests into themed subtests
